Extract floor template functions into a floors Jinja2 extension
